attempting to rebuild with the deposit closets staggered
Unfortunately I realized that there would only have been space for a 1.4 meter door and usable space in the deposit closet if they're stacked ontop of each other (because 1 meter of the 2.4 meter floor-to-ceiling space is occuped by the 0.8 meter-high drum closet below the user's toilet seat.

Rather than make some users feel claustrophobic, it make sense to stagger the deposit closets so they have a full 2.4 meters of space.

To do this without fucking-up the rise/run on the ramp, we'll need the ramp to fork -- one path will continue level to the next deposit closet and the other path will go outward and continue sloping upward. The forks should happen on a level landing, so we should keep all forks on the edges and the deposit closets on the edges