progress on landing heights for the building sunk 1.1 meters underground.
I decided to have a bit of a basement such that the first deposit closet is at ground level. With our slope and the placement of the deposit closet, that means one floor will be 1.1 meters below ground. That's not an even divider of the ceiling height of 2.4 meters, but if we just look at the run distance between landings, we can calculate the height of each landing by multiplying the run distance by 0.1 (which is our target slope).

So if it runs the entire 7 meters, then it'll increase 0.7 meters. If it only runs 6 meters, it'll increase 0.6 meters. In some cases (such as when we need a leval landing to go around the corner *and* a level landing to enter a door *and* when this all happens on a 6-meter side, the run is only 5 meters of sloped ramp. In this case the landing should only be 0.5 meters higher on the z-index than the previous one.