added first window. also realized some height issues.
While I think this design masterfully handles the slope of the walkways and the spacing of deposit closets at an offset higher than the main entrance (so they're elevated above the 200L drum that they're "depositing" into, they might actually be too high.

I chose 1.2 meters because that's where the seismic band should be placed in confined masonry.

But that means that's just where the user's feet will be. In a squat toilet, that means they're hovering 1.2 meters above the ground. The standard 200L drum is "just under 880 mm tall"

 * https://en.wikipedia.org/wiki/Drum_(container)

That means a squat toilet would be 1200-880 = 0.4 meters above the drum.

But if it's a sitting-toilet height (as opposed to a squat toilet) then it's 431-483 cm off the ground. Then that's something between 0.751 - 0.803 meters between the user and the top of the barrel.

So if their, err, angle of attack is not perpendicular, I guess there's more room for mess? Or maybe it doesn't matter. Maybe it's actually better to have more space, so bigger containers could be used in the future? Maybe?