changed the footprint to 8 meters x 8 meters
I'm changing the size to be 8 meters x 8 meters. That allows for a 2 x 2 meter space in the center for an elevator (1x1 isn't big enough for 1.2m pallets plus buffer for the elevator structure itself) plus 3x4 meter "changing places" rooms.

This is huge, but I think it's best if we want it to be accessible and easy to move pallet-sized boxes full of shit between floors.