reduced structure height to 5 floors
after reading Mexican, Colombian, and EU building codes for confined masonry structures, it became clear that 5 stories is the max. There are some others, eg India, that allow 6-stories -- but probably it would be best to be conservative here and stick with the most restrictive building codes, especially because many consumers of our designs will not be professional civil engineers or builder.

According to the Eurocode, these "storey" restrictions do not restrict below-ground floors (eg basements), so 5 storeys above ground + 1 storey below ground should be OK.

* page 203 https://www.confinedmasonry.org/wp-content/uploads/2009/09/Eurocode-8-1-Earthquakes-general.pdf

I'm also going to have to drastically reduce the size of the windows. According to the Peruvian building code, "window openings must be less than half the length of the wall".

* page 15 https://www.confinedmasonry.org/wp-content/uploads/2009/09/Peru_Booklet-masonry-v03.pdf)

I also made the basement depth an entry in the spreadsheet, so it can by dynamically referenced by other elements (such as doors and windows)