adjusted location of first maintenance door
Rather than calling it an "entrance door", I'm remaining it to "maintenance door" -- because this is the door that people will enter to do maintenance (as opposed to the doors that people will enter to use the toilets).

I set the wall thickness to be 200 mm in the spreadsheet and moved the door over this distance in the X direction. Then sunk the door -1 * the basement depth, so it becomes our first "maintenance door" that's (currently) 1.2 meters below ground-level, where the first set of toilets will be located.