added platform for the maintenance door 0
In this commit, I created a sketch (which references the lines of the main floor sketch) and adds a flat "platform" that goes at the base of the maintenance door 0. I made the "width" of the platform set to a value in the spreadsheet called "ramp width". Currently the ramp (and platforms in front of doors along the path of the wrap-around ramp) is set to 1 meter

I also added the extruded "part" of the platform, to make it 3D. Currently it's extruded to 200 mm, but this is also defined in the spreadsheet (as ramp_thickness)

Next, I'll create an array of these ramps. Then I'll create platforms for the other two doors.

Per my notes in todo.txt, I'll put two platforms for two doors to WCs on the same side of the second "wind" around the ramp. This will make it so that one of the toilets is 0.6 meters above the maintenance floor below. That's not tall enough, but if I make it a sitting toilet that's 0.4 meters above the floor of the WC, then that's exactly 1 meter. Then the second toilet will be elevated another ~0.6 meters, making the floor of this WC 0.6+0.6 = 1.2 meters above the maintance floor. Probably I'd want this one to be a squat toilet, so that it's only 0.2 meters above the top of the 1-meter-high barrel below it.