added an array of platforms for the entrance to the maintenance door
Next I need to figure out where the platforms will go for the two WCs on the next wind of the ramp.

Since we changed to 8 meters x 8 meters for the external wall footprint of the structure, it gets us closer to meeting the US ADA slope.

US ADA ramp slope is 1" of rise for every 12" of run. We're not using inches, but the ratio still applies. 1/12 = 0.083.

Australia is less strict. They have 1/8 = 0.125.

Previously I was targeting somewhere between these -- where rise/run = 0.083 - 0.125. From my notes, we chose the 6x7 meter structure specifically so we could meet that requirement. We had 5+5+6+6-1 = 2.4/21 = 0.114. So it was just *barely* sufficient for wheelchairs per Australia's code, but quite a bit steeper than the US would allow. Let's see if we can beat the US requirements now that we have larger walls @ 8x8 meter.

So, if the only flat landings that we had were on the corner (that's the 9th meter beyond the wall), then we'd have a total of 8+8+8+8 = 32 meters of sloped ramp. 2.4/32 = 0.075. Cool, but we have to have some landings. Each landing is 1 meter large. We need three of them (one for the maintenance door, one for the first WC, and one for the second WC) -- unless we do the "bypass" of the ramp like we had in the last design (we did this to keep the shitter's height above the maintenance floor more reasonable, but the ramp length run is another plus).

 0 landings = 2.4/32 = 0.075
 1 landing  = 2.4/31 = 0.0774
 2 landings = 2.4/30 = 0.08
 3 landings = 2.4/29 = 0.0828
 4 landings = 2.4/28 = 0.0857

Therefore, it looks like 3 landings satisfies the US requirements (<0.083), but 4 landings breaks it. Perfect!

The next question is: at what height should we put each of these landings?

We already finished the first landing (for the maintenance doors).

The second landing for the 1st WC is going to have a sloped ramp of length 8 meters (the length of the wall) minus 1 meter (the length of the landing for the maintenance door) = 7 meters going to it. If our slope is 2.4/29 = 0.082758621, then this landing would ideally be 0.082758621×7 = 0.579310345 meters above the first landing.

The third landing for the 2nd WC is going to have a sloped ramp of length 8 meters (the length of the wall) minus 2 meters (the combined length of both landings for both WCs, which are on the same side of the structure) = 6 meters going to it. If our slope is 2.4/29 = 0.082758621, then this landing would ideally be 0.082758621×6 = 0.496551724 meters above the second landing.

Now I'll try to place these landings and see if my calculations make sense.