I made the platform for the first WC. The floor of the first WC appears to be 0.57229 meters above the floor of the basement. To fit a 1-meter-high crate under it, that means the toilet needs to sit 1-0.57229 = 0.42771 meters off the ground.
A 40 cm toilet is a bit low, but I think it would work for people of all sizes. A 60 cm toilet would work too, but it probably wouldn't work for people of all sizes (at least, not with their feet on the ground). 50 cm seems ideal, and anywhere between 40 cm - 50 cm should work. So I think this (43 cm) is great.