added extrusions for ramp corners
the final corner that connects to the maintenance door on floor #2 appears to be perfectly level with the platform for the maintenance door #2, so it appears that our calculations using the slope of 2.4/29 = 0.082758621 to determine the z-index placement for each of these ramp platforms was correct.

And that means we've achieved a slope within the spec of the US ADA. Hooray!