added first segment of ramp surface
I used the Draft tool to draw a 3D wire between the platforms; this is the floor of our sloped ramp. Looks like I'm supposed to connect the vertices on the bottom of the platforms (as opposed to the top) since they're extruding upwards.

I'll try to extrude this in the next commit. If it looks good, then I'll repeat for the other 3x sides. Then comes an array to complete the ramp up to the last maintenance door on the top floor of the structure