added rest of basement ramp
this commit added a sketch, array and extrusions that completes the ramp down to the basement maintenance door. Unfortunately, I thought that I could extrude a subset of the sketch, but it looks like I can't. So there's a couple floating platforms that are out-of-place.

In my next commit, I'll have to divide the sketch into two pieces to eliminate those floating platforms mid-ground