finished ramp to basement
This commit finishes the ramp to basement, such that it turns 180 degrees so as not to intersect with the entrance to our WCs on the ground level.

Unfortunately, I just went from the first turn to the ground level, without calculating the slope. The result is that the slope is 0.117 -- which does satisfy Australia's requirements, but falls short of meeting the US requirements. For reference, our slop for the rest of the ramp around the structure is 0.082758621.

We could probably get away with keeping it as-is, because wheelchair users shouldn't need to go into the maintenance area of the basement. But we will be hauling a lot of extremely heavy hand truck loads of containers of shit out of there, so it might be wise to keep the slope low to make this easier when shifting the shit around between floors (eventually we may have an elevator in the air well in the middle of the structure, but that likely won't happen for many who build this structure; we shouldn't expect it).

So, we could reduce the slope to match the rest of the ramp and either have it just keep doing 180s out in the same direction away from the structure -- or we could turn 90 degrees and head around the back of the structure.

Another thing I just realized is that the ramp is 1 meter wide. This is probably fine for wheelchairs, but if the intention is to use it to shift 1.2 meter x 0.8 meter (EU-sized) pallets between floors, then we might have problems with the turning radius of a hydraulic hand truck. I'll need to research how to calculate aisles in warehouses for forklifts/hand trucks to see if this needs to be widened or if it's OK.