added doors for all three WCs on the ground floor
This commit places the doors for the entrance to the WCs on the ground floor.

In reality, only one of the doors is actually at "ground level". I decided to make the second Changing Place WC 200 mm below ground level, so that the door would be a proper 1985 mm high. Note that both of these CP doors are 1.2 meters wide. That's because the min clearance for CP doors is 1 meter, and the additional 200 mm gives us 100 mm on either side for the door framing. We probably won't need that much, but it's not an issue to have a bit more than 1 meters clearance for wheelchairs entering the CP WC

The third WC is ~0.620 meters below ground level. It's not intended to be an accessible toilet, but it's already accessible by the ramp that goes to the basement. This is probably going to be the most frequently-used WC by able-bodied folks, so I might also add stairs that go down to it. I also made an array out of this door, since it's the same placement for all the WCs located directly above it

Next I should put platforms at the base of each of these three WCs, and ramps between them from the ground level