added ramp between the two CP platforms
This commit adds a ramp between the two Changing Places WCs. The first one is 200 mm below ground-level, and the second one is at ground-level.

The distance between the two platforms is 1.6 meters. It changes 0.2 meters, so that's rise/run = 0.2/1.6 = 0.125, which too steep to meet US ADA standards, but it does Australia's maximum slope for wheelchair ramps. It's not ideal, but I can't really change the placement of the door due to other constraints. It doesn't matter too much, though because we're going to add a proper ramp perpendicualr to this one that goes down to the first CP WC. This ramp was just additional. I figure it would be nice for some folks to have this short path between the two rooms, so they don't have to go out and back if they discover that the WC is occupied when they reach the door. Worst-case, we can put some barrier to prevent wheelchairs from using this slightly-more-dangerous path, but I really don't think that will be necessary.

Next I should build the ADA-compliant ramp that goes to this CP1 WC, and also build another (possibly not ADA-compliant ramp) that goes to the WC all the way to the left