added ADA-compliant ramp down to CP1
This commit adds a 2.5 meter long ramp down to the entrance to the first Changing Place WC. 0.2/2.5 = 0.08, which (unlike the steeper ramp going between the CP WCs) is ADA-compliant. It's kinda turning into a maze of ramps, but this is all at or below ground level, so they can just be made by cutting & tamping the dirt. Easy.