added floor slabs and light well
I realized last night that the first changing place cuts into our light well.

This commit adds the floor slabs for the maintenance floors and the light well, which we'll pocket out of those slabs. That should help us "see" issues like this in the future.

Next I need to actually pocket out the towering light well solid out of the floor slabs. Then I need to change the size of the first Changing Places WC from 3x4 to 4x3. That means we only have 1 meter width (but plenty of length) for the last WC. It'll be tight, but it should work.