cut light well out of floor slabs
In this commit, I converted the light well to a padded part and then used the "Make a cut of two shapes" button of the "Part" workbench to cut the light well body as a void out of the slabs.

It's clear in this commit that the 1st WC cuts into the path of the light well.

Next I need to resize the CP1 WC.