changed ramp width from 1 meter to 2 meters
I spent most the day today trying to figure out if the ramps that we have would be sufficient to use for moving 1200 mm x 800 mm EUR-pallets (~1 meter high, full of composting shit) on manual hydraulic pallet jacks between floors.

My conclusion was that the slope that we have wouldn't be a problem. Currently we have a slope of 2.4/29 = 0.082758621 slope (8.28% grade). That's just slightly better than the minimum ADA requirement of  1/12 = 0.083333333 slope (8.33% grade). Unfortunately, all the EU standards are locked-up in PDFs that you have to pay for (wtf), but it appears that 1:20 is not uncommon. That made me a bit scared, but 1:12 should be sufficient. I do think we should try to shoot for 1:20 for the two CP WCs and the first floor, however.

Anyway, this all should be good enough for moving heavy pallets. According to OSHA 1926.451(e)(5)(ii):

> No ramp or walkway shall be inclined more than a slope of one (1) vertical to three (3) horizontal". That's 1/3 = 0.3333 slope = 33.33% grade

Some warehouses advocate for a maximum ramp slope of 10% grade.

> Our safety recommendation sets the maximum material handling ramp slope of 5.7% degrees (10% grade) which calculates to a ratio of 10’ of ramp for every 1’ of rise

At 8.28% grade, we're almost 2% better than that :)

The other concern was maneuverability: do we have enough space on our paths to push/pull a fully-loaded pallet jack along the ramps? The pallet jack would only be 800 mm wide, and our paths are 1m wide (perhaps a bit less if we add rails), so that should be fine. My concern was the 180 degree turns on the ramp down to the basement.

To determine the minimum width of the ramp required for a pallet jack, what we need to determine is the Wall-to-Wall Turning Circle.

This is something we should get from the manufacturer specifications of the pallet jack (typically in a table with a column titled "outer turning radius")

A quick google for some manufacturers of pallet jacks with fork lengths = 1200 mm (the length of EUR-pallets) shows a slightly higher turning radius than the length of the pallet, ranging from 1265 to 1400 mm. An electric pallet jack (if we ever got one) is a bit longer. A forklift is even longer. But I found some forklifts with a turning radius as tight as ~1.6 meters.

More info here:

 * https://engineering.stackexchange.com/questions/60826/warehouse-aisles-and-ramp-requirements-for-pallet-hand-trucks/

Also, I saw that some barrier-free guides said you needed at least 1.5 meter landings.

So, what we have isn't enough. I could *just* increase the landings (and keep the ramps 1m wide). Or I could just make the thing future-proof and increase the width of all the ramps to 1.6 meters. I decided to round this up to 2 meters. Honestly, if volunteers are going to be hauling huge & heavy loads of shit up & down these ramps, that extra buffer will probably be very appreciated.

The ramp's width was set by a value in a spreadsheet. All I did was change that from '=1m' to '=2m'. Obviously, now what's needed is to do some cleanup.