increased maintenance door landings to 3.5 m long
This commit changes the size of the landing outside all of the maintenance doors to be 3.5 meters long.

I realized that the landings outside the maintenance doors weren't big enough. The doors are only just barely wide enough for the 800 mm wide pallets to pass through, so they must come out straight. The pallets are 1200 mm long, so that means that they'll already be at 1200 mm onto the landing before they can even *start* their turn. I learned recently (see previous commits from today) that the turning radius of manual pallet jacks are commonly 1265 mm - 1335 mm, and electric powered pallet jacks are more, around 1600 mm. So that means the *minimum* required distance for someone to make a U-turn with a EUR-pallet full-of-shit on a pallet jack, they'd need an additional 1.2 - 1.6 meters once they start their turn. But they can only start their turn when they're 1.2 meters onto the landing, and with a 2 meter landing, they'd only have 0.6 meters of space left!

I'm not going to make the ramps wider; that's unreasonable and unnecessary. And we're not going to be carrying pallets into the WCs (those get stored below the WCs and only enter/exit the structure at the maintenance doors).

Instead, I just decided to make the landings outside the maintenance doors 3.5 meters wide. That way, they have 3.5 - 1.2 = 2.3 meters of space to make a turn after they exit the door. That should suffice.