increased landing size outside maintenance doors
This commit further increases the size of the landing outside of the landing doors.

In my last commit log, I explained why I increased the length of the landing outside the maintenance door to 3.5 meters, so that I'd have enough distance to clear the edge railing when turning the pallet jack around after exiting the door.

Well, I realized that wasn't enough. It *was* enough for that one dimension, but I needed to actually draw the arc to see how wide the landing would need to be in the *other* direction to complete the turn. In this commit, I actually drew the damn circle. I used reference lines in the sketch of the landing so that the turn starts at 1200 meters away from the door (that's the length of the pallet). Then I drew a circle with a radius of 2 meters (that's a conservative turning radius for the pallet jack). Finally, I made the edge of the landing tangent to that circle. The end length was 4.5 meters.

So now the landing is 3.5 meters by 4.5 meters. That's big, but it's necessary to haul pallets full of shit in-and-out of those doors.

I also made the ramp widen gradually, so it starts at 2-meters-wide at the previous landing, and it widens to the 4.5 meter landing here.

That actually only covers when we turn right to go down the ramp when exiting the maintenance door. We should probably do the same for turning left to go up the ramp.