fixed ramp to cp1
This commit fixes the ramp to changing place 1. Since this is the one ramp that we definitely expect to be used by wheelchair users (and because the dimensions allow for it), I made it have a slope of 1:20 (instead of ~1:12). This means that, for a rise of 200 mm, we need (0.2 meter rise) / (1/20) = 4 meters of run