added stairs to WC1
In this commit, I added a (superfluous) set of stairs descending down to the WC1 (and the basement). This allows faster access to what is probably going to be the most-commonly-used toilet for folks who can take stairs.

While the wheelchair ramp follows the best-practice of 1:20 (5% grade), the stairs follows the best-practice of rise of 150 mm & going of 300 mm = 150/300 = 50% grade. The stairs didn't line-up exactly as I expected, so I'm not 100% on my math here. The WC1 is located 820.69 mm below the surface of the ground (it's a strange number that's a result of the slop of the ramp, which is the length of all sides of the building minus 3 meters for the landing), and to get from that to ground-level at a slope of 0.5 = 820.69 / 0.5 = 1641.38 mm.

So I put the platform for the top of the stairs (at ground-level) 1.641.38 meters away from the platform height for the WC1, drew a line between them, and used the Stairs tool in the Arch workbench to create stairs. You can't set the rise/run in FreeCAD -- just the number of steps.

4 steps produced a rise of 155.17 mm and a run of 584.94 mm. Too steep.

5 steps produced a rise of 124.14 mm and a run of 438.70 mm. That satisfies our requirements.