lowered entire structure 200 mm more
This commit lowered the walls and basement slab an additional 0.2 meters, which is the current setting for the thickness of the floor slabs.

This was necessary because the wc1 floor slab was only 200 mm above the basement slab. With a toilet seat height of 480 mm, we'd need at lest 520 mm of clearance between the top of the basement floor slab and the bottom of the wc1 floor slab in-order to fit the 1000 mm pallet bin underneath the toilet seat.

I got 200 mm additional by lowering the basement floor slab.

I got 200 mm additional by raising the wc1 floor slab. Now there's a 200mm step-up for these WCs. Not ideal, but also not an issue as we have two Changing Places toilets for this structure