added toilet seats to wc1
This commit created the spot where people will actually sit in wc1 when making their deposit.

Note that the material for the toilet seat (which will probably be made of wood or bamboo) is 100mm thick, and the top of it is exactly 470 mm above the floor where the person stands. This is supposed to be 480 mm from the top of the toilet seat to the floor. The extra 10mm will be for the rim of a plastic toilet seat

I made a solid block body that covers this area and cut it out of the floor underneath this toilet seat section (there will be three seats; one for defecating into, one for urinating into, and one for bum wash -- each are 500 mm x 500mm in a row). This gives us a bit over 1000 mm clearance underneath the toilet seats. The plumbing for the urine & water seats doesn't need much, but the defecation hole needs 1 meter underneath it, for the removable pallet bin.